Refactor: api 분리
