+++ src/constants/index.ts
@@ -2,7 +2,7 @@ export const Routes = {
   Home: () => "/",
   Post: (slug: string) => `/post/${slug}`,
   Category: (category?: string) =>
-    category ? `/category/${category}` : "/category",
+    category ? `/categories/${category}` : "/categories",
 } as const;
 
 export const NotionColorMap = {
